Create new migration script for market domination data
