+++ api/streaming/socket_client.py
@@ -1,5 +1,6 @@
 from streaming.socket_manager import BinanceWebsocketClient
 
+
 class SpotWebsocketStreamClient(BinanceWebsocketClient):
     ACTION_SUBSCRIBE = "SUBSCRIBE"
     ACTION_UNSUBSCRIBE = "UNSUBSCRIBE"
