+++ api/streaming/streaming_controller.py
@@ -166,13 +166,23 @@ class BbspreadsUpdater(BaseStreaming):
 
         # multiplied by 1000 to get to the same scale stop_loss
         top_spread = round_numbers(
-            abs((bb_spreads["bb_high"] - bb_spreads["bb_mid"]) / bb_spreads["bb_high"])
-            * 100,
+            (
+                abs(
+                    (bb_spreads["bb_high"] - bb_spreads["bb_mid"])
+                    / bb_spreads["bb_high"]
+                )
+                * 100
+            ),
             2,
         )
         whole_spread = round_numbers(
-            abs((bb_spreads["bb_high"] - bb_spreads["bb_low"]) / bb_spreads["bb_high"])
-            * 100,
+            (
+                abs(
+                    (bb_spreads["bb_high"] - bb_spreads["bb_low"])
+                    / bb_spreads["bb_high"]
+                )
+                * 100
+            ),
             2,
         )
         bottom_spread = round_numbers(
