+++ api/tests/test_assets.py
@@ -1,4 +1,3 @@
-from time import sleep
 from fastapi.testclient import TestClient
 import pytest
 from account.assets import Assets
@@ -119,27 +118,6 @@ def patch_store_balance(monkeypatch):
         "create_balance_series",
         lambda self, total_balance, total_estimated_fiat: None,
     )
-
-
-def test_get_market_domination(patch_database):
-
-    response = app_client.get("/account/market-domination")
-
-    # Assert the expected result
-    expected_result = {
-        "data": {
-            "dates": ["2024-02-19 18:17:40.479"],
-            "gainers_percent": [1235816253.3],
-            "losers_percent": [0.0],
-            "gainers_count": [1],
-            "losers_count": [0],
-            "total_volume": [801797585.14104],
-        },
-        "message": "Successfully retrieved market domination data.",
-        "error": 0,
-    }
-    assert response.status_code == 200
-    assert response.json() == expected_result
 
 
 def test_get_raw_balance(patch_database, patch_raw_balances):
